@@ -1,10 +1,8 @@
 package cn.com.wind.mcp.registry;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
+import org.springframework.test.context.ActiveProfiles;
 /**
  * <p>
  * Wind MCP Registry 应用测试类
@@ -13,15 +11,15 @@ import org.springframework.test.context.junit4.SpringRunner;
  * @author yangkai.shen
  * @date Created in 2018-11-08 18:10
  */
-@RunWith(SpringRunner.class)
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ActiveProfiles("test")
 public class WindMcpRegistryApplicationTests {
 
     /**
      * 测试应用上下文加载
      */
     @Test
-    public void contextLoads() {
+    void contextLoads() {
         // 测试Spring Boot应用上下文是否能正常加载
     }
 
